backend/requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
